+++ CHANGES.md
@@ -1,3 +1,7 @@
+####2.0.0
+- Support for custom classes and ID's of which children elements will be targeted by WP-Magnific.
+- Major retooling, which will wipe out previous custom styles.
+
 ####1.1.1
 - Added enqueue parameter to ensure jQuery is loaded prior to wp-magnific scripts.